+++ app/models/tag.rb
@@ -1,0 +1,6 @@
+class Tag < ActiveRecord::Base
+
+  # Model Relationships
+  belongs_to :variable
+
+end