+++ db/migrate/20130701151252_remove_status_from_dictionaries.rb
@@ -1,0 +1,9 @@
+class RemoveStatusFromDictionaries < ActiveRecord::Migration
+  def up
+    remove_column :dictionaries, :status
+  end
+
+  def down
+    add_column :dictionaries, :status, :string
+  end
+end