+++ test/controllers/dictionaries_controller_test.rb
@@ -7,16 +7,16 @@ class DictionariesControllerTest < ActionController::TestCase
     @dictionary = dictionaries(:one)
   end
 
-  test "should remove all concepts and relations" do
-    post :remove_concepts_and_relations, id: @dictionary
+  test "should remove all variables and domains" do
+    post :clean, id: @dictionary
     assert_not_nil assigns(:dictionary)
-    assert_equal 0, assigns(:dictionary).concepts.size
-    assert_equal 0, assigns(:dictionary).concept_property_concepts.size
+    assert_equal 0, assigns(:dictionary).variables.size
+    assert_equal 0, assigns(:dictionary).domains.size
     assert_redirected_to assigns(:dictionary)
   end
 
-  test "should not remove concepts and relations for invalid dictionary" do
-    post :remove_concepts_and_relations, id: -1
+  test "should not remove variables and domains for invalid dictionary" do
+    post :clean, id: -1
     assert_nil assigns(:dictionary)
     assert_redirected_to dictionaries_path
   end
@@ -34,7 +34,7 @@ class DictionariesControllerTest < ActionController::TestCase
 
   test "should create dictionary" do
     assert_difference('Dictionary.count') do
-      post :create, dictionary: { name: 'Dictionary Three', description: "Dictionary Test", status: 'active', visible: true }
+      post :create, dictionary: { name: 'Dictionary Three', description: "Dictionary Test", visible: true }
     end
 
     assert_redirected_to dictionary_path(assigns(:dictionary))
@@ -42,7 +42,7 @@ class DictionariesControllerTest < ActionController::TestCase
 
   test "should not create dictionary with blank name" do
     assert_difference('Dictionary.count', 0) do
-      post :create, dictionary: { name: '', description: "Dictionary Test", status: 'active', visible: true }
+      post :create, dictionary: { name: '', description: "Dictionary Test", visible: true }
     end
 
     assert_not_nil assigns(:dictionary)
@@ -52,34 +52,25 @@ class DictionariesControllerTest < ActionController::TestCase
 
   test "should create a dictionary from a CSV file" do
     assert_difference('Dictionary.count') do
-      post :create, dictionary: { name: 'Dictionary Three', description: "Dictionary Test", status: 'active', visible: true },
-                    dictionary_file: fixture_file_upload('../../test/support/dictionaries/tiny_dictionary.csv')
+      post :create, dictionary: { name: 'Dictionary Three', description: "Dictionary Test", visible: true },
+                    variables_file: fixture_file_upload('../../test/support/dictionaries/tiny_dictionary_variables.csv'),
+                    domains_file:   fixture_file_upload('../../test/support/dictionaries/tiny_dictionary_domains.csv')
     end
 
     assert_not_nil assigns(:dictionary)
-    assert_equal 7, assigns(:dictionary).concepts.size
-    assert_redirected_to dictionary_path(assigns(:dictionary))
-  end
-
-  test "should create a dictionary from a CSV file using test_01.csv" do
-    assert_difference('Dictionary.count') do
-      post :create, dictionary: { name: 'Dictionary Test 01 CSV', description: "Dictionary Test", status: 'active', visible: true },
-                    dictionary_file: fixture_file_upload('../../test/support/dictionaries/test_01.csv')
-    end
-
-    assert_not_nil assigns(:dictionary)
-    assert_equal 9, assigns(:dictionary).concepts.size
+    assert_equal 5, assigns(:dictionary).variables.size
+    assert_equal 2, assigns(:dictionary).domains.size
     assert_redirected_to dictionary_path(assigns(:dictionary))
   end
 
   test "should create empty dictionary with invalid file format" do
     assert_difference('Dictionary.count') do
-      post :create, dictionary: { name: 'Dictionary Four', description: "Dictionary Test", status: 'active', visible: true },
-                    dictionary_file: fixture_file_upload('../../test/support/dictionaries/tiny_sleep.sql')
+      post :create, dictionary: { name: 'Dictionary Four', description: "Dictionary Test", visible: true },
+                    variables_file: fixture_file_upload('../../test/support/dictionaries/tiny_sleep.sql')
     end
 
     assert_not_nil assigns(:dictionary)
-    assert_equal "Unsupported dictionary file format!", flash[:alert]
+    assert_equal "Unsupported variables file format!", flash[:alert]
     assert_redirected_to dictionary_path(assigns(:dictionary))
   end
 
@@ -101,7 +92,7 @@ class DictionariesControllerTest < ActionController::TestCase
   end
 
   test "should display errors on update dictionary with blank name" do
-    put :update, id: @dictionary, dictionary: { name: '', description: 'Description', status: 'active', visible: true }
+    put :update, id: @dictionary, dictionary: { name: '', description: 'Description', visible: true }
     assert_not_nil assigns(:dictionary)
     assert assigns(:dictionary).errors.size > 0
     assert_template 'edit'
@@ -115,19 +106,21 @@ class DictionariesControllerTest < ActionController::TestCase
 
   test "should update dictionary from a CSV file" do
     put :update, id: @dictionary, dictionary: @dictionary.attributes,
-                 dictionary_file: fixture_file_upload('../../test/support/dictionaries/tiny_dictionary.csv')
+                 variables_file: fixture_file_upload('../../test/support/dictionaries/tiny_dictionary_variables.csv'),
+                 domains_file:   fixture_file_upload('../../test/support/dictionaries/tiny_dictionary_domains.csv')
 
     assert_not_nil assigns(:dictionary)
-    assert_equal 7, assigns(:dictionary).concepts.size
+    assert_equal 5, assigns(:dictionary).variables.size
+    assert_equal 2, assigns(:dictionary).domains.size
     assert_redirected_to dictionary_path(assigns(:dictionary))
   end
 
   test "should update dictionary with invalid file format" do
     put :update, id: @dictionary, dictionary: @dictionary.attributes,
-                 dictionary_file: fixture_file_upload('../../test/support/dictionaries/tiny_sleep.sql')
+                 variables_file: fixture_file_upload('../../test/support/dictionaries/tiny_sleep.sql')
 
     assert_not_nil assigns(:dictionary)
-    assert_equal "Unsupported dictionary file format!", flash[:alert]
+    assert_equal "Unsupported variables file format!", flash[:alert]
     assert_redirected_to dictionary_path(assigns(:dictionary))
   end
 
@@ -144,18 +137,6 @@ class DictionariesControllerTest < ActionController::TestCase
     assert_difference('Dictionary.current.count', 0) do
       delete :destroy, id: -1
     end
-    assert_nil assigns(:dictionary)
-    assert_redirected_to dictionaries_path
-  end
-
-  test "should export to csv" do
-    get :export_to_csv, id: @dictionary
-    assert_not_nil assigns(:dictionary)
-    assert_response :success
-  end
-
-  test "should not export invalid dictionary to csv" do
-    get :export_to_csv, id: -1
     assert_nil assigns(:dictionary)
     assert_redirected_to dictionaries_path
   end